Refactor Working Group History tab
