+++ packages/ui/src/app/pages/WorkingGroups/WorkingGroup/components/OpeningsHistory.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import { OpeningsPagination } from '@/working-groups/components/OpeningsList'
+
+export const OpeningsHistory = ({ groupId }: { groupId: string | undefined }) => (
+  <OpeningsPagination groupId={groupId} type="past" />
+)
